+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Nedenfor ses dato for release og beskrivelse af opgaver som er implementeret.
 * Tilføj conditions til "Mere"-element
 * Tilføjet Api request handler logging.
 * Tilføjet get organized handler logging.
+* Add handler id to errors.
 * Opdaterede til [OS2Forms CPR Lookup
   1.7.0](https://github.com/itk-dev/os2forms_cpr_lookup/releases/tag/1.7.0).
 * Opdaterede til [OS2Forms CVR Lookup